--- deal/configs/default.yaml
@@ -1,28 +1,0 @@
-
-
-input_data:
-  filenames: [ traj-selection.xyz ]
-
-flare_calc:
-  gp: SGP_Wrapper
-  kernels:
-  - name: NormalizedDotProduct
-    sigma: 2
-    power: 2
-  descriptors: 
-  - name: B2
-    nmax: 8
-    lmax: 3
-    cutoff_function: cosine
-    radial_basis: chebyshev
-  cutoff: 4.5
-  #species: [] # if not specified, they will desumed from the first frame
-
-otf:
-  std_tolerance_factor: -0.1
-  update_style: threshold # or add_n
-  update_threshold: 0.05
-  max_atoms_added: -1 # or n
-  force_only: true
-  write_model: 1
-  wall_time: 86000
